Restructure + refactor frontend types
