+++ src/frontend/lib/PanelNode.ts
