+++ src/frontend/lib/api/CommandRegistryApi.ts
@@ -1,6 +1,6 @@
 import type { ElectronWindowApi } from "electron-affinity/window";
 import { commandStore } from "../stores/CommandStore";
-import type { ICommand } from "../../shared/types/index";
+import type { ICommand } from "../../../shared/types/index";
 
 export class CommandRegistryApi implements ElectronWindowApi<CommandRegistryApi> {
   registryChanged(results: ICommand[]) {
