+++ src/frontend/lib/stores/ShortcutStore.ts
