+++ src/frontend/main.ts
@@ -1,5 +1,5 @@
 import "svelte";
-import App from "./App.svelte";
+import App from "./ui/base/App.svelte";
 
 const app = new App({
   target: document.body,
