+++ src/frontend/ui/base/App.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
-  import { init } from "./Blix";
-  import { blixStore } from "./stores/BlixStore";
+  import { init } from "../../lib/Blix";
+  import { blixStore } from "../../lib/stores/BlixStore";
+  import Navbar from "./Navbar.svelte";
 
   import Layout from "./layout/Layout.svelte";
-  import Navbar from "./layout/Navbar.svelte";
   import Palette from "./palette/Palette.svelte";
 </script>
