+++ src/frontend/ui/base/layout/Panel.svelte
@@ -8,17 +8,17 @@
 
 <script lang="ts">
   import { Pane, Splitpanes } from "svelte-splitpanes";
-  import { PanelNode, PanelGroup, PanelLeaf, type PanelType } from "./PanelNode";
   import PanelBlip from "./PanelBlip.svelte";
   import { createEventDispatcher } from "svelte";
   import TileSelector from "./TileSelector.svelte";
 
-  import Graph from "./tiles/Graph.svelte";
-  import Media from "./tiles/Media.svelte";
-  import Blank from "./tiles/Blank.svelte";
-  import Debug from "./tiles/Debug.svelte";
-  import WebView from "./tiles/WebView.svelte";
-  import ShortcutSettings from "./tiles/ShortcutSettings.svelte";
+  import Graph from "../../tiles/Graph.svelte";
+  import Media from "../../tiles/Media.svelte";
+  import Blank from "../../tiles/Blank.svelte";
+  import Debug from "../../tiles/Debug.svelte";
+  import WebView from "../../tiles/WebView.svelte";
+  import ShortcutSettings from "../../tiles/ShortcutSettings.svelte";
+  import { PanelGroup, PanelLeaf, type PanelNode, type PanelType } from "lib/PanelNode";
 
   // import { scale } from "svelte/transition";
