+++ src/frontend/ui/base/layout/PanelBlip.svelte
