+++ src/frontend/ui/base/layout/PanelTabs.svelte
