+++ src/frontend/ui/base/palette/PaletteItem.svelte
