+++ src/frontend/ui/tiles/Blank.svelte
