+++ src/frontend/ui/utils/Image.svelte
