+++ src/frontend/ui/utils/graph/NodeUIFragment.svelte
