+++ src/global.d.ts
@@ -1,4 +1,4 @@
-import type { MainApis } from "./frontend/api/apiInitializer";
+import type { MainApis } from "./frontend/lib/api/apiInitializer";
 
 declare global {
   interface Window {
